+++ wireview/utils.py
@@ -40,6 +40,7 @@ from asgiref.sync import async_to_sync
 from channels.db import database_sync_to_async as db
 from django.utils.datastructures import MultiValueDict
 
+from . import telemetry
 from .core.transport import get_broker
 
 log = logging.getLogger("wireview")
@@ -81,7 +82,11 @@ def send_to(channel: str | None, type: str, **kwargs: t.Any) -> None:
         **kwargs: Additional keyword arguments to include in the message.
     """
     if channel:
-        async_to_sync(get_broker().publish)(channel, dict(type=type, channel=channel, **kwargs))
+        broker = get_broker()
+        message = dict(type=type, channel=channel, **kwargs)
+        with telemetry.span(telemetry.broadcast_published, sender=broker.__class__, topic=channel) as span:
+            span.measure(message)
+            async_to_sync(broker.publish)(channel, message)
 
 
 @on_commit
@@ -109,7 +114,11 @@ async def asend_to(channel: str | None, type: str, **kwargs: t.Any) -> None:
         **kwargs: Additional keyword arguments to include in the message.
     """
     if channel:
-        await get_broker().publish(channel, dict(type=type, channel=channel, **kwargs))
+        broker = get_broker()
+        message = dict(type=type, channel=channel, **kwargs)
+        with telemetry.span(telemetry.broadcast_published, sender=broker.__class__, topic=channel) as span:
+            span.measure(message)
+            await broker.publish(channel, message)
 
 
 async def asend_notification(channel: str, **kwargs: t.Any) -> None: